+++ consensus/Cargo.toml
@@ -26,5 +26,6 @@ parking_lot = "0.11"
 rand = "0.8"
 
 [dev-dependencies]
+aleph-bft-mock = { path = "../mock" }
 env_logger = "0.9"
 tokio = { version = "1.6.1", features = ["macros", "rt", "rt-multi-thread"] }
