+++ consensus/src/testing/byzantine.rs
@@ -1,9 +1,12 @@
 use crate::{
     member::UnitMessage::NewUnit,
     network::NetworkDataInner::Units,
-    testing::mock::{
-        configure_network, init_log, spawn_honest_member, AlertHook, Data, Hash64, Hasher64,
-        KeyBox, Network, NetworkData, Spawner,
+    testing::{
+        init_log,
+        mock::{
+            configure_network, spawn_honest_member, AlertHook, Data, Hash64, Hasher64, KeyBox,
+            Network, NetworkData, Spawner,
+        },
     },
     units::{ControlHash, FullUnit, PreUnit, SignedUnit, UnitCoord},
     Hasher, Network as NetworkT, NetworkData as NetworkDataT, NodeCount, NodeIndex, NodeMap,
