+++ consensus/src/testing/mock.rs
@@ -1,8 +1,12 @@
+use crate::{
+    exponential_slowdown, run_session, units::UnitCoord, Config, DataProvider as DataProviderT,
+    DelayConfig, FinalizationHandler as FinalizationHandlerT, Hasher, Index, KeyBox as KeyBoxT,
+    MultiKeychain as MultiKeychainT, Network as NetworkT, NodeCount, NodeIndex,
+    PartialMultisignature as PartialMultisignatureT, Receiver, Recipient, Round, Sender,
+    SpawnHandle, TaskHandle,
+};
 use async_trait::async_trait;
 use codec::{Decode, Encode};
-use log::{debug, error};
-use parking_lot::Mutex;
-
 use futures::{
     channel::{
         mpsc::{unbounded, UnboundedReceiver, UnboundedSender},
@@ -10,7 +14,8 @@ use futures::{
     },
     Future, StreamExt,
 };
-
+use log::{debug, error};
+use parking_lot::Mutex;
 use std::{
     cell::RefCell,
     collections::{hash_map::DefaultHasher, HashMap},
@@ -20,24 +25,6 @@ use std::{
     task::{Context, Poll},
     time::Duration,
 };
-
-use crate::{
-    exponential_slowdown, run_session,
-    runway::{NotificationIn, NotificationOut},
-    units::{Unit, UnitCoord},
-    Config, DataProvider as DataProviderT, DelayConfig,
-    FinalizationHandler as FinalizationHandlerT, Hasher, Index, KeyBox as KeyBoxT,
-    MultiKeychain as MultiKeychainT, Network as NetworkT, NodeCount, NodeIndex,
-    PartialMultisignature as PartialMultisignatureT, Receiver, Recipient, Round, Sender,
-    SpawnHandle, TaskHandle,
-};
-
-pub fn init_log() {
-    let _ = env_logger::builder()
-        .filter_level(log::LevelFilter::max())
-        .is_test(true)
-        .try_init();
-}
 
 pub fn gen_config(node_ix: NodeIndex, n_members: NodeCount) -> Config {
     let delay_config = DelayConfig {
@@ -57,6 +44,35 @@ pub fn gen_config(node_ix: NodeIndex, n_members: NodeCount) -> Config {
     }
 }
 
+pub fn spawn_honest_member(
+    spawner: Spawner,
+    node_index: NodeIndex,
+    n_members: NodeCount,
+    network: impl 'static + NetworkT<NetworkData>,
+) -> (UnboundedReceiver<Data>, oneshot::Sender<()>, TaskHandle) {
+    let data_provider = DataProvider::new(node_index);
+
+    let (finalization_provider, finalization_rx) = FinalizationHandler::new();
+    let config = gen_config(node_index, n_members);
+    let (exit_tx, exit_rx) = oneshot::channel();
+    let spawner_inner = spawner.clone();
+    let member_task = async move {
+        let keybox = KeyBox::new(n_members, node_index);
+        run_session(
+            config,
+            network,
+            data_provider,
+            finalization_provider,
+            keybox,
+            spawner_inner.clone(),
+            exit_rx,
+        )
+        .await
+    };
+    let handle = spawner.spawn_essential("member", member_task);
+    (finalization_rx, exit_tx, handle)
+}
+
 // A hasher from the standard library that hashes to u64, should be enough to
 // avoid collisions in testing.
 #[derive(PartialEq, Eq, Hash, Clone, Debug)]
@@ -73,130 +89,6 @@ impl Hasher for Hasher64 {
 }
 
 pub(crate) type Hash64 = <Hasher64 as Hasher>::Hash;
-
-// This struct allows to create a Hub to interconnect several instances of the Consensus engine, without
-// requiring the Member wrapper. The Hub notifies all connected instances about newly created units and
-// is able to answer unit requests as well. WrongControlHashes are not supported, which means that this
-// Hub should be used to run simple tests in honest scenarios only.
-// Usage: 1) create an instance using new(n_members), 2) connect all n_members instances, 0, 1, 2, ..., n_members - 1.
-// 3) run the HonestHub instance as a Future.
-pub(crate) struct HonestHub {
-    n_members: usize,
-    ntfct_out_rxs: HashMap<NodeIndex, UnboundedReceiver<NotificationOut<Hasher64>>>,
-    ntfct_in_txs: HashMap<NodeIndex, UnboundedSender<NotificationIn<Hasher64>>>,
-    units_by_coord: HashMap<UnitCoord, Unit<Hasher64>>,
-}
-
-impl HonestHub {
-    pub(crate) fn new(n_members: usize) -> Self {
-        HonestHub {
-            n_members,
-            ntfct_out_rxs: HashMap::new(),
-            ntfct_in_txs: HashMap::new(),
-            units_by_coord: HashMap::new(),
-        }
-    }
-
-    pub(crate) fn connect(
-        &mut self,
-        node_ix: NodeIndex,
-    ) -> (
-        UnboundedSender<NotificationOut<Hasher64>>,
-        UnboundedReceiver<NotificationIn<Hasher64>>,
-    ) {
-        let (tx_in, rx_in) = unbounded();
-        let (tx_out, rx_out) = unbounded();
-        self.ntfct_in_txs.insert(node_ix, tx_in);
-        self.ntfct_out_rxs.insert(node_ix, rx_out);
-        (tx_out, rx_in)
-    }
-
-    fn send_to_all(&mut self, ntfct: NotificationIn<Hasher64>) {
-        assert!(
-            self.ntfct_in_txs.len() == self.n_members,
-            "Must connect to all nodes before running the hub."
-        );
-        for (_ix, tx) in self.ntfct_in_txs.iter() {
-            tx.unbounded_send(ntfct.clone()).ok();
-        }
-    }
-
-    fn send_to_node(&mut self, node_ix: NodeIndex, ntfct: NotificationIn<Hasher64>) {
-        let tx = self
-            .ntfct_in_txs
-            .get(&node_ix)
-            .expect("Must connect to all nodes before running the hub.");
-        tx.unbounded_send(ntfct).expect("Channel should be open");
-    }
-
-    fn on_notification(&mut self, node_ix: NodeIndex, ntfct: NotificationOut<Hasher64>) {
-        match ntfct {
-            NotificationOut::CreatedPreUnit(pu, _parent_hashes) => {
-                let hash = pu.using_encoded(Hasher64::hash);
-                let u = Unit::new(pu, hash);
-                let coord = UnitCoord::new(u.round(), u.creator());
-                self.units_by_coord.insert(coord, u.clone());
-                self.send_to_all(NotificationIn::NewUnits(vec![u]));
-            }
-            NotificationOut::MissingUnits(coords) => {
-                let mut response_units = Vec::new();
-                for coord in coords {
-                    match self.units_by_coord.get(&coord) {
-                        Some(unit) => {
-                            response_units.push(unit.clone());
-                        }
-                        None => {
-                            panic!("Unit requested that the hub does not know.");
-                        }
-                    }
-                }
-                let ntfct = NotificationIn::NewUnits(response_units);
-                self.send_to_node(node_ix, ntfct);
-            }
-            NotificationOut::WrongControlHash(_u_hash) => {
-                panic!("No support for forks in testing.");
-            }
-            NotificationOut::AddedToDag(_u_hash, _hashes) => {
-                // Safe to ignore in testing.
-                // Normally this is used in Member to answer parents requests.
-            }
-        }
-    }
-}
-
-impl Future for HonestHub {
-    type Output = ();
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let mut ready_ixs: Vec<NodeIndex> = Vec::new();
-        let mut buffer = Vec::new();
-        for (ix, rx) in self.ntfct_out_rxs.iter_mut() {
-            loop {
-                match rx.poll_next_unpin(cx) {
-                    Poll::Ready(Some(ntfct)) => {
-                        buffer.push((*ix, ntfct));
-                    }
-                    Poll::Ready(None) => {
-                        ready_ixs.push(*ix);
-                        break;
-                    }
-                    Poll::Pending => {
-                        break;
-                    }
-                }
-            }
-        }
-        for (ix, ntfct) in buffer {
-            self.on_notification(ix, ntfct);
-        }
-        for ix in ready_ixs {
-            self.ntfct_out_rxs.remove(&ix);
-        }
-        if self.ntfct_out_rxs.is_empty() {
-            return Poll::Ready(());
-        }
-        Poll::Pending
-    }
-}
 
 #[derive(Clone)]
 pub struct Spawner {}
@@ -381,6 +273,20 @@ impl Future for UnreliableRouter {
     }
 }
 
+pub fn configure_network(
+    n_members: NodeCount,
+    reliability: f64,
+) -> (UnreliableRouter, Vec<Network>) {
+    let peer_list = n_members.into_iterator().collect();
+    let mut router = UnreliableRouter::new(peer_list, reliability);
+    let mut networks = Vec::new();
+    for ix in n_members.into_iterator() {
+        let network = router.connect_peer(ix);
+        networks.push(network);
+    }
+    (router, networks)
+}
+
 #[async_trait]
 pub trait NetworkHook: Send {
     /// This must complete during a single poll - the current implementation
@@ -559,74 +465,4 @@ impl MultiKeychainT for KeyBox {
     fn is_complete(&self, _: &[u8], partial: &Self::PartialMultisignature) -> bool {
         (self.count * 2) / 3 < NodeCount(partial.signed_by.len())
     }
-}
-
-pub(crate) async fn run_honest_member<N: 'static + NetworkT<NetworkData>>(
-    config: Config,
-    network: N,
-    data_provider: DataProvider,
-    finalization_provider: FinalizationHandler,
-    keybox: KeyBox,
-    spawn_handle: Spawner,
-    exit: oneshot::Receiver<()>,
-) {
-    run_session(
-        config,
-        network,
-        data_provider,
-        finalization_provider,
-        keybox,
-        spawn_handle,
-        exit,
-    )
-    .await
-}
-
-pub fn configure_network(
-    n_members: NodeCount,
-    reliability: f64,
-) -> (UnreliableRouter, Vec<Network>) {
-    let peer_list = n_members.into_iterator().collect();
-    let mut router = UnreliableRouter::new(peer_list, reliability);
-    let mut networks = Vec::new();
-    for ix in n_members.into_iterator() {
-        let network = router.connect_peer(ix);
-        networks.push(network);
-    }
-    (router, networks)
-}
-
-pub fn spawn_honest_member(
-    spawner: Spawner,
-    node_index: NodeIndex,
-    n_members: NodeCount,
-    network: impl 'static + NetworkT<NetworkData>,
-) -> (UnboundedReceiver<Data>, oneshot::Sender<()>, TaskHandle) {
-    let data_provider = DataProvider::new(node_index);
-
-    let (finalization_provider, finalization_rx) = FinalizationHandler::new();
-    let config = gen_config(node_index, n_members);
-    let (exit_tx, exit_rx) = oneshot::channel();
-    let spawner_inner = spawner.clone();
-    let member_task = async move {
-        let keybox = KeyBox::new(n_members, node_index);
-        run_honest_member(
-            config,
-            network,
-            data_provider,
-            finalization_provider,
-            keybox,
-            spawner_inner.clone(),
-            exit_rx,
-        )
-        .await
-    };
-    let handle = spawner.spawn_essential("member", member_task);
-    (finalization_rx, exit_tx, handle)
-}
-
-pub fn complete_oneshot<T: std::fmt::Debug>(t: T) -> oneshot::Receiver<T> {
-    let (tx, rx) = oneshot::channel();
-    tx.send(t).unwrap();
-    rx
 }
